♻️ refactor: move heterogeneous session error types to @lobechat/types
Removes the @lobechat/electron-client-ipc barrel edge from web importers
by relocating HeterogeneousAgentSessionErrorCode, HeterogeneousAgentRateLimitInfo,
and HeterogeneousAgentSessionError to a package with no Electron dependency.
